\docType{package}
\name{gglc}
\alias{gglc}
\alias{gglc-package}
\title{gglc: A simple package to produce ggplots from Lorenz curves}
\description{
  Making Lorenz curves work in ggplot2 isn't hard. It's
  just kind of a pain in the ass. This package is here to
  improve that.
}
\details{
  Plus making multiple curves at once is fun.
}

